Osm code

+++ src/main/java/game/SoloGame.java
@@ -3,7 +3,7 @@ package game;
 import java.util.List;
 
 public class SoloGame extends AbstractGame {
-    public SoloGame(int id, GameMap gameMap, List<AbstractPlayer> players, float webSpeed, float mobileMaxSpeed, List<Dwarf> dwarfs, Integer timeToEnd) {
+    public SoloGame(int id, GameMap gameMap, List<AbstractPlayer> players, double webSpeed, double mobileMaxSpeed, List<Dwarf> dwarfs, Integer timeToEnd) {
         super(id, gameMap, players, webSpeed, mobileMaxSpeed, dwarfs, timeToEnd);
     }
 }
